requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
selenium==4.15.0
webdriver-manager==4.0.1
gunicorn==21.2.0